apps/03-food-simple/backend/python/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
qdrant-client==1.7.3
elasticsearch==8.12.1
psycopg2-binary==2.9.9
weaviate-client==4.5.0
numpy==1.26.4
python-multipart==0.0.9
httpx==0.27.0
aiohttp==3.9.3
asyncpg==0.29.0
google-generativeai==0.3.2 